arm64: add explicit cast to int in field extraction
The lack of an explicit typecast in a ternary conditional produces the
following compiler message with -Werror, so add an explicit cast to int
because the value ends up being used as an int anyway:

include/asm/cpufeature.h: In function 'cpuid_feature_extract_field':
include/asm/cpufeature.h:161:55: error: signed and unsigned type in conditional expression [-Werror=sign-compare]

Bug 1853519

Reviewed-on: http://git-master/r/1327687
GVS: Gerrit_Virtual_Submit

+++ arch/arm64/include/asm/cpufeature.h
@@ -159,7 +159,7 @@ cpuid_feature_extract_field(u64 features, int field, bool sign)
 {
 	return (sign) ?
 		cpuid_feature_extract_signed_field(features, field) :
-		cpuid_feature_extract_unsigned_field(features, field);
+		(int)cpuid_feature_extract_unsigned_field(features, field);
 }
 
 static inline s64 arm64_ftr_value(const struct arm64_ftr_bits *ftrp, u64 val)
